requirements.txt
fastapi>=0.115,<1.0
granian>=2.7,<3.0
httpx[http2]>=0.28,<1.0
pydantic>=2.9,<3.0
websocket-client>=1.8,<2.0
pytest>=8.3,<9.0
